mcp-geo2neo/pyproject.toml
[project]
name = "mcp-geo2neo"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastmcp>=2.0.0",
  "pydantic>=2.7.0",
  "orjson>=3.10.0",
  "neo4j>=5.20.0",
  "python-dotenv>=1.0.1",
]


